backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
groq
python-multipart
supabase
googlemaps
python-dotenv